+++ src/backend/core/urls.py
@@ -56,4 +56,5 @@ urlpatterns = [
         ),
     ),
     path(f"api/{settings.API_VERSION}/config/", viewsets.ConfigView.as_view()),
+    path(f"api/{settings.API_VERSION}/footer/", viewsets.FooterView.as_view()),
 ]