+++ src/backend/impress/settings.py
@@ -410,7 +410,14 @@ class Base(Configuration):
     FRONTEND_THEME = values.Value(
         None, environ_name="FRONTEND_THEME", environ_prefix=None
     )
-
+    FRONTEND_URL_JSON_FOOTER = values.Value(
+        None, environ_name="FRONTEND_URL_JSON_FOOTER", environ_prefix=None
+    )
+    FRONTEND_FOOTER_VIEW_CACHE_TIMEOUT = values.Value(
+        60 * 60 * 24,
+        environ_name="FRONTEND_FOOTER_VIEW_CACHE_TIMEOUT",
+        environ_prefix=None,
+    )
     FRONTEND_CSS_URL = values.Value(
         None, environ_name="FRONTEND_CSS_URL", environ_prefix=None
     )